Move Metric classes into separate class. Rename MetricsResult -> FindDeterioratedMetrics
